feat: Add parent command ability * Fixes #22
Adds ability to start a Cypress chain where the previous subject will be the body element. This is helpful for using a piped function that is meant for an element to not scope to some container element.

This allows the folowing:
```ts
const getChildren = $el => $el.children()

cy.pipe(getChildren)
  .should('have.length', 3)
```